Split up various different API client "flavors"
